+++ tests/handler/test_fastapi.py
@@ -8,37 +8,21 @@ from fastapi import FastAPI
 from fastapi.exceptions import RequestValidationError
 from starlette.exceptions import HTTPException
 
-from web_error import error
-from web_error.cors import CorsConfiguration
-from web_error.handler import fastapi
+from fastapi_problem import error
+from fastapi_problem.cors import CorsConfiguration
+from fastapi_problem.handler import fastapi
 
 
-class SomethingWrongError(error.ServerException):
+class SomethingWrongError(error.ServerProblem):
     title = "This is an error."
 
 
-class CustomUnhandledException(error.ServerException):
+class CustomUnhandledException(error.ServerProblem):
     title = "Unhandled exception occurred."
 
 
-class CustomValidationError(error.HttpCodeException):
+class CustomValidationError(error.StatusProblem):
     status = 422
-    title = "Request validation error."
-
-
-class ALegacyError(error.ServerException):
-    title = "This is an error."
-    code = "E123"
-
-
-class LegacyUnhandledException(error.ServerException):
-    code = "E000"
-    title = "Unhandled exception occurred."
-
-
-class LegacyValidationError(error.HttpCodeException):
-    status = 422
-    code = "E001"
     title = "Request validation error."
 
 
@@ -53,190 +37,6 @@ def cors():
 
 
 class TestExceptionHandler:
-    def test_unexpected_error_replaces_code(self):
-        logger = mock.Mock()
-
-        request = mock.Mock()
-        exc = Exception("Something went bad")
-
-        eh = fastapi.generate_handler(
-            logger=logger,
-            unhandled_wrappers={
-                "default": CustomUnhandledException,
-            },
-        )
-        response = eh(request, exc)
-
-        assert response.status_code == http.HTTPStatus.INTERNAL_SERVER_ERROR
-        assert json.loads(response.body) == {
-            "title": "Unhandled exception occurred.",
-            "details": "Something went bad",
-            "type": "custom-unhandled-exception",
-            "status": 500,
-        }
-        assert logger.exception.call_args == mock.call(
-            "Unhandled exception occurred.",
-            exc_info=(type(exc), exc, None),
-        )
-
-    @pytest.mark.backwards_compat()
-    def test_unexpected_error_replaced_legacy(self):
-        logger = mock.Mock()
-
-        request = mock.Mock()
-        exc = Exception("Something went bad")
-
-        eh = fastapi.generate_handler(
-            logger=logger,
-            unhandled_wrappers={
-                "default": LegacyUnhandledException,
-            },
-            legacy=True,
-        )
-        response = eh(request, exc)
-
-        assert response.status_code == http.HTTPStatus.INTERNAL_SERVER_ERROR
-        assert json.loads(response.body) == {
-            "message": "Unhandled exception occurred.",
-            "debug_message": "Something went bad",
-            "code": "E000",
-        }
-        assert logger.exception.call_args == mock.call(
-            "Unhandled exception occurred.",
-            exc_info=(type(exc), exc, None),
-        )
-
-    def test_strip_debug(self):
-        request = mock.Mock()
-        exc = Exception("Something went bad")
-
-        eh = fastapi.generate_handler(
-            strip_debug=True,
-            unhandled_wrappers={
-                "default": CustomUnhandledException,
-            },
-        )
-        response = eh(request, exc)
-
-        assert response.status_code == http.HTTPStatus.INTERNAL_SERVER_ERROR
-        assert json.loads(response.body) == {
-            "title": "Unhandled exception occurred.",
-            "type": "custom-unhandled-exception",
-            "status": 500,
-        }
-
-    def test_strip_debug_with_code(self):
-        request = mock.Mock()
-        exc = Exception("Something went bad")
-
-        eh = fastapi.generate_handler(
-            strip_debug=False,
-            strip_debug_codes=[500],
-            unhandled_wrappers={
-                "default": CustomUnhandledException,
-            },
-        )
-        response = eh(request, exc)
-
-        assert response.status_code == http.HTTPStatus.INTERNAL_SERVER_ERROR
-        assert json.loads(response.body) == {
-            "title": "Unhandled exception occurred.",
-            "type": "custom-unhandled-exception",
-            "status": 500,
-        }
-
-    def test_strip_debug_with_allowed_code(self):
-        request = mock.Mock()
-        exc = SomethingWrongError("something bad")
-
-        eh = fastapi.generate_handler(
-            strip_debug=False,
-            strip_debug_codes=[400],
-            unhandled_wrappers={
-                "default": CustomUnhandledException,
-            },
-        )
-        response = eh(request, exc)
-
-        assert response.status_code == http.HTTPStatus.INTERNAL_SERVER_ERROR
-        assert json.loads(response.body) == {
-            "title": "This is an error.",
-            "type": "something-wrong",
-            "details": "something bad",
-            "status": 500,
-        }
-
-    @pytest.mark.backwards_compat()
-    def test_strip_debug_legacy(self):
-        request = mock.Mock()
-        exc = Exception("Something went bad")
-
-        eh = fastapi.generate_handler(
-            strip_debug=True,
-            unhandled_wrappers={
-                "default": LegacyUnhandledException,
-            },
-            legacy=True,
-        )
-        response = eh(request, exc)
-
-        assert response.status_code == http.HTTPStatus.INTERNAL_SERVER_ERROR
-        assert json.loads(response.body) == {
-            "message": "Unhandled exception occurred.",
-            "code": "E000",
-        }
-
-    def test_unexpected_error(self):
-        logger = mock.Mock()
-
-        request = mock.Mock()
-        exc = Exception("Something went bad")
-
-        eh = fastapi.generate_handler(logger=logger)
-        response = eh(request, exc)
-
-        assert response.status_code == http.HTTPStatus.INTERNAL_SERVER_ERROR
-        assert json.loads(response.body) == {
-            "title": "Unhandled exception occurred.",
-            "details": "Something went bad",
-            "type": "unhandled-exception",
-            "status": 500,
-        }
-        assert logger.exception.call_args == mock.call(
-            "Unhandled exception occurred.",
-            exc_info=(type(exc), exc, None),
-        )
-
-    def test_known_error(self):
-        request = mock.Mock()
-        exc = SomethingWrongError("something bad")
-
-        eh = fastapi.generate_handler()
-        response = eh(request, exc)
-
-        assert response.status_code == http.HTTPStatus.INTERNAL_SERVER_ERROR
-        assert json.loads(response.body) == {
-            "title": "This is an error.",
-            "details": "something bad",
-            "type": "something-wrong",
-            "status": 500,
-        }
-
-    @pytest.mark.backwards_compat()
-    def test_known_error_legacy(self):
-        request = mock.Mock()
-        exc = ALegacyError("something bad")
-
-        eh = fastapi.generate_handler(legacy=True)
-        response = eh(request, exc)
-
-        assert response.status_code == http.HTTPStatus.INTERNAL_SERVER_ERROR
-        assert json.loads(response.body) == {
-            "message": "This is an error.",
-            "debug_message": "something bad",
-            "code": "E123",
-        }
-
     def test_fastapi_error(self):
         request = mock.Mock()
         exc = RequestValidationError([])
@@ -255,61 +55,6 @@ class TestExceptionHandler:
             "type": "custom-validation",
             "status": 422,
         }
-
-    @pytest.mark.backwards_compat()
-    def test_fastapi_error_legacy(self):
-        request = mock.Mock()
-        exc = RequestValidationError([])
-
-        eh = fastapi.generate_handler(
-            unhandled_wrappers={
-                "422": LegacyValidationError,
-            },
-            legacy=True,
-        )
-        response = eh(request, exc)
-
-        assert response.status_code == http.HTTPStatus.UNPROCESSABLE_ENTITY
-        assert json.loads(response.body) == {
-            "message": "Request validation error.",
-            "debug_message": [],
-            "code": "E001",
-        }
-
-    def test_starlette_error(self):
-        request = mock.Mock()
-        exc = HTTPException(http.HTTPStatus.NOT_FOUND, "something bad")
-
-        eh = fastapi.generate_handler()
-        response = eh(request, exc)
-
-        assert response.status_code == http.HTTPStatus.NOT_FOUND
-        assert json.loads(response.body) == {
-            "title": "Not Found",
-            "details": exc.detail,
-            "type": "http-not-found",
-            "status": 404,
-        }
-
-    def test_starlette_error_with_headers(self):
-        request = mock.Mock()
-        exc = HTTPException(
-            status_code=http.HTTPStatus.UNAUTHORIZED,
-            detail="Incorrect username or password",
-            headers={"WWW-Authenticate": "Basic"},
-        )
-
-        eh = fastapi.generate_handler()
-        response = eh(request, exc)
-
-        assert response.status_code == http.HTTPStatus.UNAUTHORIZED
-        assert json.loads(response.body) == {
-            "title": "Unauthorized",
-            "details": exc.detail,
-            "type": "http-unauthorized",
-            "status": 401,
-        }
-        assert response.headers["www-authenticate"] == "Basic"
 
     def test_error_with_origin(self, cors):
         request = mock.Mock(headers={"origin": "localhost"})