+++ tests/test_error.py
@@ -1,6 +1,6 @@
 import pytest
 
-from web_error import error
+from fastapi_problem import error
 
 
 class NotFoundError(error.NotFoundException):
@@ -42,61 +42,3 @@ def test_marshal(exc, type_):
         "details": "details",
         "status": e.status,
     }
-
-
-@pytest.mark.backwards_compat()
-def test_legacy_attributes_new_error():
-    e = NotFoundError("details")
-
-    assert e.message == e.title
-    assert e.code is None
-    assert e.debug_message == e.details
-
-
-@pytest.mark.backwards_compat()
-def test_legacy_attributes():
-    e = ALegacyError("details")
-
-    assert e.message == e.title
-    assert e.code == e.type
-    assert e.debug_message == e.details
-
-
-@pytest.mark.backwards_compat()
-def test_marshal_legacy():
-    e = ALegacyError("debug_message")
-
-    assert e.marshal(legacy=True) == {
-        "code": "E500",
-        "message": "a 500 message",
-        "debug_message": "debug_message",
-    }
-
-
-@pytest.mark.backwards_compat()
-def test_init_legacy():
-    e = error.HttpException(
-        message="Unable to connect to service.",
-        debug_message="debug_message",
-        code="E000",
-        status=400,
-    )
-
-    assert e.marshal() == {
-        "type": "E000",
-        "title": "Unable to connect to service.",
-        "details": "debug_message",
-        "status": 400,
-    }
-
-
-@pytest.mark.backwards_compat()
-def test_init_legacy_no_message():
-    with pytest.raises(TypeError) as e:
-        error.HttpException(
-            debug_message="debug_message",
-            code="E000",
-            status=400,
-        )
-
-    assert str(e.value) == "HttpException.__init__() missing 1 required positional argument: 'title'"